// client/src/data/city.data.ts
// Type for creating a city, without autogenerated fields
import api from "../utils/api.ts";

export interface ICitySchema {
    _id?:string;
    name: string;
    location: {
        long: number;
        lat: number;
    };
    createdAt?: Date; // Optional, auto-generated by Mongoose
}


const getCitiesData = async ()=>{
    return await api.get('/cities')
}
const showCityData = async (id:string)=>{
    return await api.get(`/cities/${id}`)
}

export {
    getCitiesData,
    showCityData,
}